fixes from PR feedback
