+++ src/ui/components/LinkNewTab.tsx
@@ -1,0 +1,10 @@
+import { Link } from '@mui/material';
+
+export function LinkNewTab(props: { href: string; children: React.ReactNode }) {
+  const { href, children, ...rest } = props;
+  return (
+    <Link {...rest} href={href} target="_blank" rel="noopener">
+      {children}
+    </Link>
+  );
+}
